Divide code on its module
