+++ src/components/Button.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Button = ({ text }) => (
+  <button id="button" type="button">
+    { text }
+  </button>
+);
+
+Button.propTypes = {
+  text: PropTypes.string.isRequired,
+};
+
+export default Button;
