+++ src/index.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/App';
+import './index.css';
+
 
 ReactDOM.render(
   <React.StrictMode>
